refactor plugin placements
