+++ src/main/java/dk/kb/datahandler/oai/plugins/Plugin.java
@@ -1,4 +1,4 @@
-package dk.kb.datahandler.oai;
+package dk.kb.datahandler.oai.plugins;
 
 public interface Plugin {
     void apply();
