+++ src/main/java/dk/kb/datahandler/oai/plugins/PreservicaManifestationPlugin.java
@@ -1,0 +1,8 @@
+package dk.kb.datahandler.oai.plugins;
+
+public class PreservicaManifestationPlugin  implements Plugin{
+    @Override
+    public void apply() {
+
+    }
+}
